Update to Go 1.23 Read secrets from environment
