--- .travis.yml
@@ -1,19 +1,0 @@
-language: go
-
-sudo: false
-
-go:
-  - 1.1
-  - 1.2
-  - 1.3
-  - 1.4
-  - 1.5
-  - 1.6
-  - 1.7
-  - tip
-
-install:
-  - go get github.com/stretchr/testify/assert
-
-script:
-  - go test -v ./...
