+++ config/config.env
@@ -1,0 +1,1 @@
+API="https://api.test.nordnet.se/next/2"
