+++ feed/feed.go
@@ -1,6 +1,6 @@
 /*
-	Contains everything related to the public and private feeds
-	More information available on https://api.test.nordnet.se/next/2/api-docs/docs/feeds
+Contains everything related to the public and private feeds
+More information available on https://api.test.nordnet.se/next/2/api-docs/docs/feeds
 */
 package feed
 
@@ -9,48 +9,6 @@ import (
 	"encoding/json"
 	"io"
 )
-
-// Used in the UnmarshalJSON implementations on PrivateFeed and PublicFeed
-var (
-	heartbeatType     = "heartbeat"
-	orderType         = "order"
-	tradeType         = "trade"
-	priceType         = "price"
-	depthType         = "depth"
-	tradingStatusType = "trading_status"
-	indicatorType     = "indicator"
-	newsType          = "news"
-)
-
-// Used when sending feed commands
-type FeedCmd struct {
-	Cmd  string      `json:"cmd"`
-	Args interface{} `json:"args"`
-}
-
-// Arguments for sending the login command
-type LoginArgs struct {
-	SessionKey string      `json:"session_key"`
-	GetState   interface{} `json:"get_state,omitempty"`
-}
-
-// Arguments for getting orders and trades when logging in
-type GetState struct {
-	DeletedOrders bool  `json:"deleted_orders"`
-	Days          int64 `json:"days,omitempty"`
-}
-
-// Used for receiving messages
-type FeedMsg struct {
-	Type string      `json:"type"`
-	Data interface{} `json:"data"`
-}
-
-// Used in UnmarshalJSON overrides
-type rawMsg struct {
-	Type string          `json:"type"`
-	Data json.RawMessage `json:"data"`
-}
 
 // Represents the feed connection
 type Feed struct {
@@ -83,4 +41,34 @@ func (f *Feed) Close() error {
 // Send the login command with the specified session key
 func (f *Feed) Login(session string, getState interface{}) error {
 	return f.Write(&FeedCmd{Cmd: "login", Args: &LoginArgs{SessionKey: session, GetState: getState}})
+}
+
+// Used when sending feed commands
+type FeedCmd struct {
+	Cmd  string      `json:"cmd"`
+	Args interface{} `json:"args"`
+}
+
+// Arguments for sending the login command
+type LoginArgs struct {
+	SessionKey string      `json:"session_key"`
+	GetState   interface{} `json:"get_state,omitempty"`
+}
+
+// Arguments for getting orders and trades when logging in
+type GetState struct {
+	DeletedOrders bool  `json:"deleted_orders"`
+	Days          int64 `json:"days,omitempty"`
+}
+
+// Used for receiving messages
+type FeedMsg struct {
+	Type string      `json:"type"`
+	Data interface{} `json:"data"`
+}
+
+// Used in UnmarshalJSON overrides
+type rawMsg struct {
+	Type string          `json:"type"`
+	Data json.RawMessage `json:"data"`
 }
